Emulation loop

+++ include/chip8.hpp
@@ -29,13 +29,13 @@ public:
     uint32_t* getVideo();
     uint8_t* getKeypad();
     uint16_t getIndexRegister();
-    uint8_t getMemoryAt(uint8_t index);
+    uint8_t getMemoryAt(uint16_t index);
     uint8_t getDelayTimer();
     uint8_t getSoundTimer();
     uint8_t getRandomByte();
 
     void setIndexRegister(uint16_t value);
-    void writeMemory(uint8_t index, uint8_t value);
+    void writeMemory(uint16_t index, uint8_t value);
     void setDelayTimer(uint8_t value);
     void setSoundTimer(uint8_t value);
     void setKeypad(int index, uint8_t value);
